requirements.txt
mcp-server-fastmcp>=0.1.0
groundx>=0.1.0
python-dotenv>=1.0.0 